Add files via upload
The core objective of this study is to thoroughly examine the dataset of life expectancy and numerous health indicators for 193 nations from the years 2000 to 2015. The dataset contains 22 columns and 2938 rows, which correspond to 22 predictive variables of interest. It was compiled from the Global Health Observatory (GHO) under the WHO and supplemented with economic data from the United Nations.